my-first-ai-chatbot/requirements.txt
langchain
langchain-groq
python-dotenv
gradio